- Refactored package structure, - Enhanced JWT request filtering by adding improved exception handling for malformed, expired, and empty JWT tokens, - Updated authentication flow, - General code improvement by adding comments to clarify the purpose of JWT request filtering
