+++ src/main/java/com/Japkutija/veterinarybackend/veterinary/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package com.Japkutija.veterinarybackend.veterinary.config;
 
-import com.Japkutija.veterinarybackend.veterinary.security.CustomUserDetailsService;
-import com.Japkutija.veterinarybackend.veterinary.security.JwtRequestFilter;
+import com.Japkutija.veterinarybackend.veterinary.service.impl.CustomUserDetailsService;
+import com.Japkutija.veterinarybackend.veterinary.security.filters.JwtRequestFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,7 +9,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
