+++ src/main/java/com/Japkutija/veterinarybackend/veterinary/security/filters/JwtRequestFilter.java
@@ -1,7 +1,9 @@
-package com.Japkutija.veterinarybackend.veterinary.security;
+package com.Japkutija.veterinarybackend.veterinary.security.filters;
 
+import com.Japkutija.veterinarybackend.veterinary.service.impl.CustomUserDetailsService;
 import com.Japkutija.veterinarybackend.veterinary.security.util.JwtUtil;
 import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.MalformedJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -11,7 +13,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -24,6 +25,17 @@ public class JwtRequestFilter extends OncePerRequestFilter {
     private final JwtUtil jwtUtil;
     private final CustomUserDetailsService userDetailsService;
 
+    /**
+     * Filters each request to validate the JWT token and set the authentication in the security context.
+     * If a valid token is found, it sets the user's authentication details in the security context.
+     * If the token is expired, an error is logged.
+     *
+     * @param request  the HTTP request
+     * @param response the HTTP response
+     * @param chain    the filter chain
+     * @throws ServletException if an error occurs during the filtering process
+     * @throws IOException      if an I/O error occurs during the filtering process
+     */
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
             throws ServletException, IOException {
@@ -38,10 +50,15 @@ public class JwtRequestFilter extends OncePerRequestFilter {
                 username = jwtUtil.extractUsername(jwt);
             } catch (ExpiredJwtException e) {
                 log.error("JWT Token has expired");
+            } catch (MalformedJwtException e) {
+                log.error("JWT Token is malformed");
+            } catch (IllegalArgumentException e) {
+                log.error("JWT Token is null or empty");
             }
         }
 
         // Validate token and set authentication
+        // The reason we are checking for getAuthentication is to avoid re-authenticating the user
         if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
             var userDetails = this.userDetailsService.loadUserByUsername(username);
             if (jwtUtil.validateToken(jwt, userDetails.getUsername())) {
@@ -51,6 +68,7 @@ public class JwtRequestFilter extends OncePerRequestFilter {
             }
         }
 
+        // Continue the filter chain, pass the request and response to the next filter
         chain.doFilter(request, response);
     }
 }
